Initial commit.
This commit introduces the specmonkey link collector module.
This is a python module that can scan files in
a directory structure for links. For performance reasons, the actual scanning algorithm is implemented in rust.

For testing purposes, a simple unfiltered list output is available by calling
```sh
python -m specmonkey simple-list <path> <output-file>
```